refactor: move network management commands into one folder
