+++ packages/zwave-js/src/lib/controller/Controller.ts
@@ -127,6 +127,52 @@ import {
 	SetRFReceiveModeResponse,
 } from "../serialapi/misc/SetRFReceiveModeMessages";
 import {
+	AddNodeDSKToNetworkRequest,
+	AddNodeStatus,
+	AddNodeToNetworkRequest,
+	AddNodeToNetworkRequestStatusReport,
+	AddNodeType,
+	EnableSmartStartListenRequest,
+} from "../serialapi/network-mgmt/AddNodeToNetworkRequest";
+import { AssignReturnRouteRequest } from "../serialapi/network-mgmt/AssignReturnRouteMessages";
+import { AssignSUCReturnRouteRequest } from "../serialapi/network-mgmt/AssignSUCReturnRouteMessages";
+import { DeleteReturnRouteRequest } from "../serialapi/network-mgmt/DeleteReturnRouteMessages";
+import { DeleteSUCReturnRouteRequest } from "../serialapi/network-mgmt/DeleteSUCReturnRouteMessages";
+import {
+	GetSUCNodeIdRequest,
+	GetSUCNodeIdResponse,
+} from "../serialapi/network-mgmt/GetSUCNodeIdMessages";
+import {
+	IsFailedNodeRequest,
+	IsFailedNodeResponse,
+} from "../serialapi/network-mgmt/IsFailedNodeMessages";
+import {
+	RemoveFailedNodeRequest,
+	RemoveFailedNodeResponse,
+	RemoveFailedNodeStartFlags,
+	RemoveFailedNodeStatus,
+	type RemoveFailedNodeRequestStatusReport,
+} from "../serialapi/network-mgmt/RemoveFailedNodeMessages";
+import {
+	RemoveNodeFromNetworkRequest,
+	RemoveNodeFromNetworkRequestStatusReport,
+	RemoveNodeStatus,
+	RemoveNodeType,
+} from "../serialapi/network-mgmt/RemoveNodeFromNetworkRequest";
+import {
+	ReplaceFailedNodeRequest,
+	ReplaceFailedNodeRequestStatusReport,
+	ReplaceFailedNodeResponse,
+	ReplaceFailedNodeStartFlags,
+	ReplaceFailedNodeStatus,
+} from "../serialapi/network-mgmt/ReplaceFailedNodeRequest";
+import {
+	NodeNeighborUpdateStatus,
+	RequestNodeNeighborUpdateReport,
+	RequestNodeNeighborUpdateRequest,
+} from "../serialapi/network-mgmt/RequestNodeNeighborUpdateMessages";
+import { SetSUCNodeIdRequest } from "../serialapi/network-mgmt/SetSUCNodeIDMessages";
+import {
 	ExtNVMReadLongBufferRequest,
 	ExtNVMReadLongBufferResponse,
 } from "../serialapi/nvm/ExtNVMReadLongBufferMessages";
@@ -157,21 +203,9 @@ import {
 	NVMOperationsWriteRequest,
 } from "../serialapi/nvm/NVMOperationsMessages";
 import {
-	AddNodeDSKToNetworkRequest,
-	AddNodeStatus,
-	AddNodeToNetworkRequest,
-	AddNodeToNetworkRequestStatusReport,
-	AddNodeType,
-	EnableSmartStartListenRequest,
-} from "./AddNodeToNetworkRequest";
-import { AssignReturnRouteRequest } from "./AssignReturnRouteMessages";
-import { AssignSUCReturnRouteRequest } from "./AssignSUCReturnRouteMessages";
-import {
 	ControllerStatistics,
 	ControllerStatisticsHost,
 } from "./ControllerStatistics";
-import { DeleteReturnRouteRequest } from "./DeleteReturnRouteMessages";
-import { DeleteSUCReturnRouteRequest } from "./DeleteSUCReturnRouteMessages";
 import { minFeatureVersions, ZWaveFeature } from "./Features";
 import {
 	GetControllerCapabilitiesRequest,
@@ -197,10 +231,6 @@ import {
 	GetSerialApiInitDataRequest,
 	GetSerialApiInitDataResponse,
 } from "./GetSerialApiInitDataMessages";
-import {
-	GetSUCNodeIdRequest,
-	GetSUCNodeIdResponse,
-} from "./GetSUCNodeIdMessages";
 import { HardResetRequest } from "./HardResetRequest";
 import {
 	InclusionOptions,
@@ -214,39 +244,9 @@ import {
 	SmartStartProvisioningEntry,
 } from "./Inclusion";
 import {
-	IsFailedNodeRequest,
-	IsFailedNodeResponse,
-} from "./IsFailedNodeMessages";
-import {
-	RemoveFailedNodeRequest,
-	RemoveFailedNodeRequestStatusReport,
-	RemoveFailedNodeResponse,
-	RemoveFailedNodeStartFlags,
-	RemoveFailedNodeStatus,
-} from "./RemoveFailedNodeMessages";
-import {
-	RemoveNodeFromNetworkRequest,
-	RemoveNodeFromNetworkRequestStatusReport,
-	RemoveNodeStatus,
-	RemoveNodeType,
-} from "./RemoveNodeFromNetworkRequest";
-import {
-	ReplaceFailedNodeRequest,
-	ReplaceFailedNodeRequestStatusReport,
-	ReplaceFailedNodeResponse,
-	ReplaceFailedNodeStartFlags,
-	ReplaceFailedNodeStatus,
-} from "./ReplaceFailedNodeRequest";
-import {
-	NodeNeighborUpdateStatus,
-	RequestNodeNeighborUpdateReport,
-	RequestNodeNeighborUpdateRequest,
-} from "./RequestNodeNeighborUpdateMessages";
-import {
 	SetSerialApiTimeoutsRequest,
 	SetSerialApiTimeoutsResponse,
 } from "./SetSerialApiTimeoutsMessages";
-import { SetSUCNodeIdRequest } from "./SetSUCNodeIDMessages";
 import { ZWaveLibraryTypes } from "./ZWaveLibraryTypes";
 import { protocolVersionToSDKVersion } from "./ZWaveSDKVersions";
 import type { RSSI } from "./_Types";
