+++ packages/zwave-js/src/lib/serialapi/network-mgmt/AddNodeToNetworkRequest.ts
@@ -3,12 +3,12 @@ import {
 	NodeType,
 	parseNodeUpdatePayload,
 } from "@zwave-js/core";
-import type { Driver } from "../driver/Driver";
+import type { Driver } from "../../driver/Driver";
 import {
 	FunctionType,
 	MessagePriority,
 	MessageType,
-} from "../message/Constants";
+} from "../../message/Constants";
 import {
 	expectedCallback,
 	gotDeserializationOptions,
@@ -18,8 +18,8 @@ import {
 	MessageOptions,
 	messageTypes,
 	priority,
-} from "../message/Message";
-import type { SuccessIndicator } from "../message/SuccessIndicator";
+} from "../../message/Message";
+import type { SuccessIndicator } from "../../message/SuccessIndicator";
 
 export enum AddNodeType {
 	Any = 1,
