+++ packages/zwave-js/src/lib/serialapi/network-mgmt/GetNodeProtocolInfoMessages._test.ts
