+++ packages/zwave-js/src/lib/serialapi/network-mgmt/ReplaceFailedNodeRequest.ts
@@ -1,9 +1,9 @@
-import type { Driver } from "../driver/Driver";
+import type { Driver } from "../../driver/Driver";
 import {
 	FunctionType,
 	MessagePriority,
 	MessageType,
-} from "../message/Constants";
+} from "../../message/Constants";
 import {
 	expectedResponse,
 	gotDeserializationOptions,
@@ -13,8 +13,8 @@ import {
 	MessageOptions,
 	messageTypes,
 	priority,
-} from "../message/Message";
-import type { SuccessIndicator } from "../message/SuccessIndicator";
+} from "../../message/Message";
+import type { SuccessIndicator } from "../../message/SuccessIndicator";
 
 export enum ReplaceFailedNodeStartFlags {
 	OK = 0,
